Stackless issue #290: bpo-36974, PEP 590 Vectorcall-protocol
Adapt Stackless Python to the new Vectorcall protocol introduced by upstream commit aacc77fb.
This changes changes the Stackless protocol to support Vectorcall functions.
It also updates code in pricklpit.c to correctly initialize wrapper types. This avoids assertion
failures introduced by upstream commit be718c3 (not yet merged).

+++ Include/cpython/slp_structs.h
@@ -317,7 +317,7 @@ PyAPI_DATA(PyTypeObject) PyChannel_Type;
  ******************************************************/
 
 #ifndef _PyStackless_TRY_STACKLESS
-PyAPI_DATA(int * const) _PyStackless__TryStacklessPtr;
+PyAPI_DATA(intptr_t * const) _PyStackless__TryStacklessPtr;
 #define _PyStackless_TRY_STACKLESS (*_PyStackless__TryStacklessPtr)
 #endif
 #ifndef STACKLESS__GETARG_ASSERT
